Update rpcmining.cpp
Fix for getnetworkhashps overflow that causes a negative hash per second value.

+++ src/rpcmining.cpp
@@ -69,7 +69,7 @@ UniValue GetNetworkHashPS(int lookup, int height)
     uint256 workDiff = pb->nChainWork - pb0->nChainWork;
     int64_t timeDiff = maxTime - minTime;
 
-    return (int64_t)(workDiff.getdouble() / timeDiff);
+    return workDiff.getdouble() / timeDiff;
 }
 
 UniValue getnetworkhashps(const UniValue& params, bool fHelp)
